optimizes the today tasks

+++ src/main/java/com/Project/Task_Manager/Service/TaskService.java
@@ -1,20 +1,14 @@
 package com.Project.Task_Manager.Service;
 
-import DTO.TaskRequest;
-import com.Project.Task_Manager.Entity.NoteEntity;
 import com.Project.Task_Manager.Repository.NoteRepository;
-import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataAccessException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.Project.Task_Manager.Entity.TaskEntity;
 import com.Project.Task_Manager.Repository.TaskRepository;
 
-import java.time.DateTimeException;
 import java.time.LocalDate;
-import java.util.Date;
 import java.util.List;
 
 @Service
@@ -27,8 +21,8 @@ public class TaskService {
     public ResponseEntity<TaskEntity> addTask(TaskEntity taskEntity) {
         LocalDate deadlineDate=taskEntity.getDeadline();
         LocalDate createdDate =taskEntity.getCreated_At();
-        boolean isValid = deadlineDate.isAfter(createdDate);
-        if(isValid){
+        boolean before = deadlineDate.isBefore(createdDate);
+        if(!before){
             return ResponseEntity.ok(taskRepository.save(taskEntity));
         }else{
             return ResponseEntity.unprocessableEntity().build();
@@ -54,11 +48,16 @@ public class TaskService {
         List<TaskEntity> list = taskRepository.findAll();
         for (TaskEntity entity: list){
             boolean deadlineFinished = entity.getDeadline().isBefore(LocalDate.now());
-            if(deadlineFinished) {
+            if(deadlineFinished && entity.isCompleted()) {
                 taskRepository.delete(entity);
             }
         }
         return taskRepository.findAll();
+    }
+    public List<TaskEntity> getAllTodayTasks(){
+        LocalDate currentDate = LocalDate.now();
+        List<TaskEntity> tasks = taskRepository.findTasksByCurrentDate(currentDate);
+        return tasks;
     }
 
 }
